+++ R/build_run_modify.r
@@ -42,7 +42,7 @@
 # methods::setClass is called during build not package source code.
 # suppress NOTE with a spurious importFrom in the namespace
 #' @importFrom stats AIC C aggregate as.formula coef complete.cases
-#' @importFrom stats confint cor cov cov.wt cov2cor df lm
+#' @importFrom stats confint cor cov cov.wt cov2cor df lm cor.test
 #' @importFrom stats logLik na.exclude na.omit pchisq pf qchisq
 #' @importFrom stats qnorm quantile residuals rnorm runif sd
 #' @importFrom stats setNames update var delete.response terms
@@ -429,7 +429,6 @@ umxModel <- function(...) {
 #'# wt   "mpg_with_wt"   "wt_with_wt"  "b1"
 #'# disp "disp_with_mpg" "b1"          "disp_with_disp"
 umxRAM <- function(model = NA, ..., data = NULL, name = NA, comparison = TRUE, setValues = TRUE, suffix = "", independent = NA, remove_unused_manifests = TRUE, showEstimates = c("none", "raw", "std", "both", "list of column names"), refModels = NULL, autoRun = getOption("umx_auto_run"), tryHard = c("no", "mxTryHard", "mxTryHardOrdinal", "mxTryHardWideSearch"), type = c('Auto', 'FIML', 'cov', 'cor', 'WLS', 'DWLS', 'ULS'), optimizer = NULL, thresholds = c("deviationBased"), verbose = FALSE) {
-	
 	dot.items = list(...) # grab all the dot items: mxPaths, etc...
 	dot.items = unlist(dot.items) # In case any dot items are lists of mxPaths, etc...
 	showEstimates = umx_default_option(showEstimates, c("none", "raw", "std", "both", "list of column names"), check = FALSE)
@@ -470,6 +469,7 @@ umxRAM <- function(model = NA, ..., data = NULL, name = NA, comparison = TRUE, s
 	}
 
 	foundNames = c()
+	defnNames = c()
 	for (thisItem in dot.items) {
 		if(!is.list(thisItem)){
 			# Sometimes we get a list, so expand everything to a list.
@@ -479,10 +479,18 @@ umxRAM <- function(model = NA, ..., data = NULL, name = NA, comparison = TRUE, s
 			thisIs = class(thisItem[[i]])[1]
 			if(thisIs == "MxPath"){
 				foundNames = append(foundNames, c(thisItem[[i]]$from, thisItem[[i]]$to))
+				tmp = namez(thisItem[[i]]$labels, "data\\.")
+				if(length(tmp) > 0){
+					defnNames = append(defnNames, namez(tmp, "data\\.(.*)", replacement= "\\1"))
+				}
 			} else {
 				if(thisIs == "MxThreshold"){
 					# MxThreshold detected
 				} else {
+					tmp = namez(thisItem[[i]]$labels, "data\\.")
+					if(length(tmp) > 0){
+						defnNames = append(defnNames, namez(tmp, "data\\.(.*)", replacement= "\\1"))
+					}
 					# TODO: umxRAM currently not checking for unsupported items.
 					# stop("I can only handle (u)mxPaths, (u)mxMatrices, mxConstraints, and mxThreshold() objects.\n",
 					# "You have given me a", class(i)[1],"\n",
@@ -504,8 +512,8 @@ umxRAM <- function(model = NA, ..., data = NULL, name = NA, comparison = TRUE, s
 	manifestVars = unique(na.omit(umx_names(data)))
 
 	# Omit NAs from found names as empty to = can generate these spuriously
-	foundNames   = unique(na.omit(foundNames))
-
+	foundNames = unique(na.omit(foundNames))
+	defnNames  = unique(na.omit(defnNames))
 	# Anything used as a path, but not found in the data (and not a key word like "one") must be a latent
 	latentVars = setdiff(foundNames, c(manifestVars, "one"))
 	nLatent = length(latentVars)
@@ -535,12 +543,11 @@ umxRAM <- function(model = NA, ..., data = NULL, name = NA, comparison = TRUE, s
 	}else{
 		unusedManifests = setdiff(manifestVars, foundNames)
 	}
-	# used = all data columns present in found and not reserved, e.g. "one"
+	# Used = all data columns present in found and not reserved, e.g. "one"
 	usedManifests = setdiff(intersect(manifestVars, foundNames), "one")
 
-
 	if(remove_unused_manifests & length(unusedManifests) > 0){
-		data = xmu_make_mxData(data = data, type = type, manifests = usedManifests)
+		data = xmu_make_mxData(data = data, type = type, manifests = c(usedManifests, defnNames))
 	} else {
 		data = xmu_make_mxData(data= data, type = type)
 		usedManifests = manifestVars
@@ -675,6 +682,7 @@ umxRAM <- function(model = NA, ..., data = NULL, name = NA, comparison = TRUE, s
 #' 
 umxSuperModel <- function(name = 'top', ..., autoRun = getOption("umx_auto_run"), tryHard = c("no", "mxTryHard", "mxTryHardOrdinal", "mxTryHardWideSearch")) {
 	dot.items = list(...) # grab all the dot items: models...	
+	umx_check(boolean.test=is.character(name), action="stop", message="You need to set the name for the supermodel with: name = 'modelName' ")
 	nModels = length(dot.items)
 	# get list of model names
 	modelNames = c()
@@ -802,8 +810,7 @@ umxModify <- function(lastFit, update = NULL, master = NULL, regex = FALSE, free
 	
 	if(is.null(update)){
 		message("You haven't asked to do anything: the parameters that are free to be dropped are:")
-		# TODO use parameters here?
-		print(umxGetParameters(lastFit))
+		print(parameters(lastFit))
 		stop()
 	}
 
@@ -1417,14 +1424,21 @@ umxGxE_window <- function(selDVs = NULL, moderator = NULL, mzData = mzData, dzDa
 #'
 #' # Things to note:
 #' 
-#' # 1. This variable has a large variance, but umx picks good starts.
+#' # 1. This variable has a large variance, and this makes solution finding very hard.
+#' # We'll scale weight to make the Optimizer's task easier.
+#'
+#' twinData = umx_scale_wide_twin_data(data = twinData, varsToScale = c("wt"), sep = "")
+#' mzData <- twinData[twinData$zygosity %in% "MZFF", ]
+#' dzData <- twinData[twinData$zygosity %in% "DZFF", ]
 #' 
 #' # 2. umxACE can figure out variable names: provide sep= "_T" and selVar = "wt" -> "wt_T1" "wt_T2"
 #' 
 #' # 3. umxACE picks the variables it needs from the data.
-#' # 4. note: the default boundDiag = 0 lower-bounds a, c, and e at 0 (prevents mirror-solutions).
-#'         # can remove this by setting boundDiag = NULL
-#' m1 = umxACE(selDVs = "wt", dzData = dzData, mzData = mzData, sep = "", boundDiag = NULL)
+#' # 4. expert user note: by default, umxACE lower-bounds a, c, and e at 0.
+#' #    This prevents mirror-solutions.
+#' #    You can remove this by setting boundDiag = NULL
+#' 
+#' m1 = umxACE(selDVs = "wt", dzData = dzData, mzData = mzData, sep = "")
 #'
 #' # MODEL MODIFICATION
 #' # We can modify this model, say testing shared environment, and see a comparison:
@@ -1436,6 +1450,7 @@ umxGxE_window <- function(selDVs = NULL, moderator = NULL, mzData = mzData, dzDa
 #' # = Bivariate height and weight model =
 #' # =====================================
 #' data(twinData)
+#' twinData = umx_scale_wide_twin_data(data = twinData, varsToScale = c("wt"), sep = "")
 #' mzData = twinData[twinData$zygosity %in% c("MZFF", "MZMM"),]
 #' dzData = twinData[twinData$zygosity %in% c("DZFF", "DZMM", "DZOS"), ]
 #' mzData = mzData[1:80,] # quicker run to keep CRAN happy
@@ -1477,6 +1492,7 @@ umxGxE_window <- function(selDVs = NULL, moderator = NULL, mzData = mzData, dzDa
 #' # = Bivariate continuous and ordinal example =
 #' # ============================================
 #' data(twinData)
+#' twinData = umx_scale_wide_twin_data(data = twinData, varsToScale = c("wt"), sep = "")
 #' # Cut BMI column to form ordinal obesity variables
 #' obesityLevels   = c('normal', 'overweight', 'obese')
 #' cutPoints       = quantile(twinData[, "bmi1"], probs = c(.5, .2), na.rm = TRUE)
@@ -1496,6 +1512,7 @@ umxGxE_window <- function(selDVs = NULL, moderator = NULL, mzData = mzData, dzDa
 #' # =======================================
 #' require(umx)
 #' data(twinData)
+#' twinData = umx_scale_wide_twin_data(data = twinData, varsToScale = c("wt"), sep = "")
 #' # Cut to form category of 20% obese subjects
 #' # and make into mxFactors (ensure ordered is TRUE, and require levels)
 #' obesityLevels   = c('normal', 'obese')
@@ -1519,6 +1536,7 @@ umxGxE_window <- function(selDVs = NULL, moderator = NULL, mzData = mzData, dzDa
 #' 
 #' require(umx)
 #' data(twinData)
+#' twinData = umx_scale_wide_twin_data(data = twinData, varsToScale = c("wt"), sep = "")
 #' selDVs = c("wt1", "wt2")
 #' mz = cov(twinData[twinData$zygosity %in%  "MZFF", selDVs], use = "complete")
 #' dz = cov(twinData[twinData$zygosity %in%  "DZFF", selDVs], use = "complete")
@@ -3046,6 +3064,10 @@ umxMatrix <- function(name = NA, type = "Full", nrow = NA, ncol = NA, free = FAL
 	if(name %in% legalMatrixTypes){
 		stop("You used ", name, "as the name of your matrix. You might be used to mxMatrix, where type comes first? But it is not a legal matrix name.")
 	}
+	if(is.numeric(type)){
+		stop("You used ", omxQuotes(type), " as the type of your matrix. You probably need to add something like type='Full' or specify nrow and ncol")
+	}
+
 	if(isTRUE(labels)){
 		setLabels = TRUE
 		labels = NA